web_app/requirements.txt
Flask>=3.0
Pillow>=10.0
requests>=2.31
numpy>=1.24
urllib3>=2.0
oss2>=2.18  # 可选：阿里云 OSS 支持
